Added cookie.

+++ modules/purplejs-http/src/main/java/io/purplejs/http/Cookie.java
@@ -14,7 +14,7 @@ public interface Cookie
 
     boolean isSecure();
 
-    long getMaxAge();
+    int getMaxAge();
 
     String getPath();
 }
